Refactor golang backend and split commands
